figured out GameConfigurationDTO object and its REST representation, built gameConfig parser, changed game configuration dto logic
