--- src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/NewGameDTO.java
@@ -1,46 +1,0 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
-
-import java.util.List;
-
-import ch.uzh.ifi.hase.soprafs24.constant.GameType;
-import ch.uzh.ifi.hase.soprafs24.constant.SupportedLanguages;
-
-public class NewGameDTO {
-
-    private String host;
-    private List<String> playerNames;
-    private GameType gameType;
-    private SupportedLanguages language;
-
-    public String getHost() {
-        return host;
-    }
-
-    public void setHost(String host) {
-        this.host = host;
-    }
-
-    public List<String> getPlayerNames() {
-        return playerNames;
-    }
-
-    public void setPlayerNames(List<String> playerNames) {
-        this.playerNames = playerNames;
-    }
-
-    public GameType getGameType() {
-        return gameType;
-    }
-
-    public void setGameType(GameType gameType) {
-        this.gameType = gameType;
-    }
-
-    public SupportedLanguages getLanguage() {
-        return language;
-    }
-
-    public void setLanguage(SupportedLanguages language) {
-        this.language = language;
-    }
-}
